requirements.txt
Django==1.9.5
South==1.0.2
requests
requests_oauthlib
mysqlclient==1.3.9
djangorestframework
Pillow==3.4.2
django-cors-headers==1.3.1